Use int64 to represent byte sizes
Adapt to https://github.com/docker/model-runner/pull/57

+++ cmd/cli/commands/df.go
@@ -46,9 +46,9 @@ func diskUsageTable(df desktop.DiskUsage) string {
 	})
 	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT)
 
-	table.Append([]string{"Models", units.HumanSize(df.ModelsDiskUsage)})
+	table.Append([]string{"Models", units.HumanSize(float64(df.ModelsDiskUsage))})
 	if df.DefaultBackendDiskUsage != 0 {
-		table.Append([]string{"Inference engine", units.HumanSize(df.DefaultBackendDiskUsage)})
+		table.Append([]string{"Inference engine", units.HumanSize(float64(df.DefaultBackendDiskUsage))})
 	}
 
 	table.Render()
